Extract raw predicate plan module
